requirements.txt
# Core dependencies
streamlit>=1.32.0
Pillow>=10.0.0

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development dependencies
black>=24.1.0  # for code formatting
isort>=5.13.0  # for import sorting
flake8>=7.0.0  # for linting
mypy>=1.8.0    # for type checking